feat: ipsec_tunnel_status check Proxy ID support
`ipsec_tunnel_status` check to accept a list of `proxy_ids` to check tunnel
status against the provided list, checks for all configured Proxy IDs if not
provided.
Returns success if any Proxy ID is in active state for the ipsec tunnel or
`require_all_active` flag can be set to require all Proxy IDs to be in active state.