feat: explicit readiness checks (#186)
`ConfigParser` updated to support `explicit_elements` that makes it possible to 
skip these elements if they are not passed in requested_config explicitly.
Meaning if the `requested_config` includes "all" keyword or if it is an all exclusive config, 
the elements in the `explicit_elements` will not be included.
`CheckFirewall` updated with a set of explicit readiness checks passed to `ConfigParser`
to skip them by default with "all" requested config.